Fix MAPL_UNDEF type and re-export cascade
- InternalConstants.F90: fix MAPL_UNDEF type from real(REAL64)
  to plain real (matches MAPL_UNDEFINED_REAL)
- Base_Base.F90: re-export MAPL_UNDEF from MAPL_Constants so
  GEOS legacy code using MAPL_BaseMod still works
- MAPL_AbstractGridFactory, AbstractRegridder, VotingRegridder:
  remove redundant 'use MAPL_BaseMod, only: MAPL_UNDEF' since
  MAPL_Constants is already used in these modules
- MAPL_LlcGridFactory, TripolarGridFactory: switch local
  use from MAPL_BaseMod to MAPL_Constants
- MAPL_ObsUtil: switch to MAPL_Constants for MAPL_UNDEF

+++ shared/Constants/InternalConstants.F90
@@ -16,7 +16,7 @@ module MAPL_InternalConstants
    real,               parameter :: MAPL_UNDEFINED_REAL32  = -huge(1._REAL32)
    real(kind=REAL64),  parameter :: MAPL_UNDEFINED_REAL64  = -huge(1._REAL64)
    real,               parameter :: MAPL_UNDEFINED_REAL    = -huge(1.)
-   real(kind=REAL64),  parameter :: MAPL_UNDEF             = MAPL_UNDEFINED_REAL
+   real,               parameter :: MAPL_UNDEF             = MAPL_UNDEFINED_REAL
    integer,            parameter :: MAPL_UNDEFINED_INTEGER = -huge(1)
    character(len=*),   parameter :: MAPL_UNDEFINED_CHAR    = '**'
